backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
dj-database-url==2.1.0
whitenoise==6.5.0
twilio==8.12.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
pydub==0.25.1
requests==2.31.0
psycopg2-binary==2.9.9  # For PostgreSQL support
SpeechRecognition==3.14.2  # Speech recognition library
pocketsphinx==5.0.4  # Offline speech recognition engine 